+++ zcash_client_sqlite/src/wallet/transact.rs
@@ -165,7 +165,10 @@ mod tests {
         block::BlockHash,
         consensus::{BlockHeight, BranchId, Parameters},
         legacy::TransparentAddress,
-        sapling::{note_encryption::try_sapling_output_recovery, prover::TxProver},
+        sapling::{
+            keys::DiversifiableFullViewingKey, note_encryption::try_sapling_output_recovery,
+            prover::TxProver,
+        },
         transaction::{components::Amount, Transaction},
         zip32::{ExtendedFullViewingKey, ExtendedSpendingKey},
     };
@@ -207,8 +210,8 @@ mod tests {
         // Add two accounts to the wallet
         let extsk0 = sapling::spending_key(&[0u8; 32], network().coin_type(), AccountId::from(0));
         let extsk1 = sapling::spending_key(&[1u8; 32], network().coin_type(), AccountId::from(1));
-        let extfvk0 = ExtendedFullViewingKey::from(&extsk0);
-        let extfvk1 = ExtendedFullViewingKey::from(&extsk1);
+        let dfvk0 = DiversifiableFullViewingKey::from(ExtendedFullViewingKey::from(&extsk0));
+        let dfvk1 = DiversifiableFullViewingKey::from(ExtendedFullViewingKey::from(&extsk1));
 
         #[cfg(feature = "transparent-inputs")]
         let ufvks = {
@@ -219,14 +222,14 @@ mod tests {
                 transparent::AccountPrivKey::from_seed(&network(), &[1u8; 32], AccountId::from(1))
                     .unwrap();
             [
-                UnifiedFullViewingKey::new(Some(tsk0.to_account_pubkey()), Some(extfvk0)).unwrap(),
-                UnifiedFullViewingKey::new(Some(tsk1.to_account_pubkey()), Some(extfvk1)).unwrap(),
+                UnifiedFullViewingKey::new(Some(tsk0.to_account_pubkey()), Some(dfvk0)).unwrap(),
+                UnifiedFullViewingKey::new(Some(tsk1.to_account_pubkey()), Some(dfvk1)).unwrap(),
             ]
         };
         #[cfg(not(feature = "transparent-inputs"))]
         let ufvks = [
-            UnifiedFullViewingKey::new(Some(extfvk0)).unwrap(),
-            UnifiedFullViewingKey::new(Some(extfvk1)).unwrap(),
+            UnifiedFullViewingKey::new(Some(dfvk0), None).unwrap(),
+            UnifiedFullViewingKey::new(Some(dfvk1), None).unwrap(),
         ];
 
         init_accounts_table(&db_data, &ufvks).unwrap();
@@ -275,12 +278,12 @@ mod tests {
 
         // Add an account to the wallet
         let extsk = sapling::spending_key(&[0u8; 32], network().coin_type(), AccountId::from(0));
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
+        let dfvk = DiversifiableFullViewingKey::from(ExtendedFullViewingKey::from(&extsk));
 
         #[cfg(feature = "transparent-inputs")]
-        let ufvk = UnifiedFullViewingKey::new(None, Some(extfvk)).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(None, Some(dfvk)).unwrap();
         #[cfg(not(feature = "transparent-inputs"))]
-        let ufvk = UnifiedFullViewingKey::new(Some(extfvk)).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(Some(dfvk)).unwrap();
         init_accounts_table(&db_data, &[ufvk]).unwrap();
         let to = extsk.default_address().1.into();
 
@@ -319,11 +322,11 @@ mod tests {
 
         // Add an account to the wallet
         let extsk = sapling::spending_key(&[0u8; 32], network().coin_type(), AccountId::from(0));
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
+        let dfvk = DiversifiableFullViewingKey::from(ExtendedFullViewingKey::from(&extsk));
         #[cfg(feature = "transparent-inputs")]
-        let ufvk = UnifiedFullViewingKey::new(None, Some(extfvk)).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(None, Some(dfvk)).unwrap();
         #[cfg(not(feature = "transparent-inputs"))]
-        let ufvk = UnifiedFullViewingKey::new(Some(extfvk)).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(Some(dfvk)).unwrap();
         init_accounts_table(&db_data, &[ufvk]).unwrap();
         let to = extsk.default_address().1.into();
 
@@ -367,11 +370,11 @@ mod tests {
 
         // Add an account to the wallet
         let extsk = sapling::spending_key(&[0u8; 32], network().coin_type(), AccountId::from(0));
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
+        let dfvk = DiversifiableFullViewingKey::from(ExtendedFullViewingKey::from(&extsk));
         #[cfg(feature = "transparent-inputs")]
-        let ufvk = UnifiedFullViewingKey::new(None, Some(extfvk.clone())).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(None, Some(dfvk.clone())).unwrap();
         #[cfg(not(feature = "transparent-inputs"))]
-        let ufvk = UnifiedFullViewingKey::new(Some(extfvk.clone())).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(Some(dfvk.clone())).unwrap();
         init_accounts_table(&db_data, &[ufvk]).unwrap();
 
         // Add funds to the wallet in a single note
@@ -379,7 +382,7 @@ mod tests {
         let (cb, _) = fake_compact_block(
             sapling_activation_height(),
             BlockHash([0; 32]),
-            extfvk.clone(),
+            &dfvk,
             value,
         );
         insert_into_cache(&db_cache, &cb);
@@ -398,12 +401,7 @@ mod tests {
         );
 
         // Add more funds to the wallet in a second note
-        let (cb, _) = fake_compact_block(
-            sapling_activation_height() + 1,
-            cb.hash(),
-            extfvk.clone(),
-            value,
-        );
+        let (cb, _) = fake_compact_block(sapling_activation_height() + 1, cb.hash(), &dfvk, value);
         insert_into_cache(&db_cache, &cb);
         scan_cached_blocks(&tests::network(), &db_cache, &mut db_write, None).unwrap();
 
@@ -446,12 +444,8 @@ mod tests {
         // Mine blocks SAPLING_ACTIVATION_HEIGHT + 2 to 9 until just before the second
         // note is verified
         for i in 2..10 {
-            let (cb, _) = fake_compact_block(
-                sapling_activation_height() + i,
-                cb.hash(),
-                extfvk.clone(),
-                value,
-            );
+            let (cb, _) =
+                fake_compact_block(sapling_activation_height() + i, cb.hash(), &dfvk, value);
             insert_into_cache(&db_cache, &cb);
         }
         scan_cached_blocks(&tests::network(), &db_cache, &mut db_write, None).unwrap();
@@ -477,8 +471,7 @@ mod tests {
         }
 
         // Mine block 11 so that the second note becomes verified
-        let (cb, _) =
-            fake_compact_block(sapling_activation_height() + 10, cb.hash(), extfvk, value);
+        let (cb, _) = fake_compact_block(sapling_activation_height() + 10, cb.hash(), &dfvk, value);
         insert_into_cache(&db_cache, &cb);
         scan_cached_blocks(&tests::network(), &db_cache, &mut db_write, None).unwrap();
 
@@ -510,11 +503,11 @@ mod tests {
 
         // Add an account to the wallet
         let extsk = sapling::spending_key(&[0u8; 32], network().coin_type(), AccountId::from(0));
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
+        let dfvk = DiversifiableFullViewingKey::from(ExtendedFullViewingKey::from(&extsk));
         #[cfg(feature = "transparent-inputs")]
-        let ufvk = UnifiedFullViewingKey::new(None, Some(extfvk.clone())).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(None, Some(dfvk.clone())).unwrap();
         #[cfg(not(feature = "transparent-inputs"))]
-        let ufvk = UnifiedFullViewingKey::new(Some(extfvk.clone())).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(Some(dfvk.clone())).unwrap();
         init_accounts_table(&db_data, &[ufvk]).unwrap();
 
         // Add funds to the wallet in a single note
@@ -522,7 +515,7 @@ mod tests {
         let (cb, _) = fake_compact_block(
             sapling_activation_height(),
             BlockHash([0; 32]),
-            extfvk,
+            &dfvk,
             value,
         );
         insert_into_cache(&db_cache, &cb);
@@ -573,7 +566,7 @@ mod tests {
             let (cb, _) = fake_compact_block(
                 sapling_activation_height() + i,
                 cb.hash(),
-                ExtendedFullViewingKey::from(&ExtendedSpendingKey::master(&[i as u8])),
+                &ExtendedFullViewingKey::from(&ExtendedSpendingKey::master(&[i as u8])).into(),
                 value,
             );
             insert_into_cache(&db_cache, &cb);
@@ -604,7 +597,7 @@ mod tests {
         let (cb, _) = fake_compact_block(
             sapling_activation_height() + 22,
             cb.hash(),
-            ExtendedFullViewingKey::from(&ExtendedSpendingKey::master(&[22])),
+            &ExtendedFullViewingKey::from(&ExtendedSpendingKey::master(&[22])).into(),
             value,
         );
         insert_into_cache(&db_cache, &cb);
@@ -639,11 +632,11 @@ mod tests {
 
         // Add an account to the wallet
         let extsk = sapling::spending_key(&[0u8; 32], network.coin_type(), AccountId::from(0));
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
+        let dfvk = DiversifiableFullViewingKey::from(ExtendedFullViewingKey::from(&extsk));
         #[cfg(feature = "transparent-inputs")]
-        let ufvk = UnifiedFullViewingKey::new(None, Some(extfvk.clone())).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(None, Some(dfvk.clone())).unwrap();
         #[cfg(not(feature = "transparent-inputs"))]
-        let ufvk = UnifiedFullViewingKey::new(Some(extfvk.clone())).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(Some(dfvk.clone())).unwrap();
         init_accounts_table(&db_data, &[ufvk]).unwrap();
 
         // Add funds to the wallet in a single note
@@ -651,7 +644,7 @@ mod tests {
         let (cb, _) = fake_compact_block(
             sapling_activation_height(),
             BlockHash([0; 32]),
-            extfvk.clone(),
+            &dfvk,
             value,
         );
         insert_into_cache(&db_cache, &cb);
@@ -708,7 +701,7 @@ mod tests {
             try_sapling_output_recovery(
                 &network,
                 sapling_activation_height(),
-                &extfvk.fvk.ovk,
+                &dfvk.fvk().ovk,
                 output,
             )
         };
@@ -725,7 +718,7 @@ mod tests {
             let (cb, _) = fake_compact_block(
                 sapling_activation_height() + i,
                 cb.hash(),
-                ExtendedFullViewingKey::from(&ExtendedSpendingKey::master(&[i as u8])),
+                &ExtendedFullViewingKey::from(&ExtendedSpendingKey::master(&[i as u8])).into(),
                 value,
             );
             insert_into_cache(&db_cache, &cb);
@@ -749,11 +742,11 @@ mod tests {
 
         // Add an account to the wallet
         let extsk = sapling::spending_key(&[0u8; 32], network().coin_type(), AccountId::from(0));
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
+        let dfvk = DiversifiableFullViewingKey::from(ExtendedFullViewingKey::from(&extsk));
         #[cfg(feature = "transparent-inputs")]
-        let ufvk = UnifiedFullViewingKey::new(None, Some(extfvk.clone())).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(None, Some(dfvk.clone())).unwrap();
         #[cfg(not(feature = "transparent-inputs"))]
-        let ufvk = UnifiedFullViewingKey::new(Some(extfvk.clone())).unwrap();
+        let ufvk = UnifiedFullViewingKey::new(Some(dfvk.clone())).unwrap();
         init_accounts_table(&db_data, &[ufvk]).unwrap();
 
         // Add funds to the wallet in a single note
@@ -761,7 +754,7 @@ mod tests {
         let (cb, _) = fake_compact_block(
             sapling_activation_height(),
             BlockHash([0; 32]),
-            extfvk,
+            &dfvk,
             value,
         );
         insert_into_cache(&db_cache, &cb);